requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-auditlog==3.1.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
everapi==0.1.1
freecurrencyapi==0.1.0
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
multidict==6.4.4
packaging==25.0
pillow==11.2.1
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.1.1
urllib3==2.4.0
yarl==1.20.1
django-prometheus
faker
numpy